Change target version from 4.7 to 4.8
Changing all uses of 4.7 in documentation and code.
