+++ subprojects/tooling-api/src/crossVersionTest/groovy/org/gradle/integtests/tooling/r48/CustomProjectsLoadedAction.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.integtests.tooling.r48;
+
+import org.gradle.tooling.BuildAction;
+import org.gradle.tooling.BuildController;
+import org.gradle.tooling.model.gradle.GradleBuild;
+
+import java.io.Serializable;
+
+public class CustomProjectsLoadedAction implements BuildAction<String>, Serializable {
+    // Plugins are not applied yet, cannot have access to custom model builders. However, basic models provided by Gradle are already available.
+
+    @Override
+    public String execute(BuildController controller) {
+        GradleBuild buildModel = controller.getBuildModel();
+        assert buildModel != null;
+        return "loading";
+    }
+}
