+++ subprojects/tooling-api/src/crossVersionTest/groovy/org/gradle/integtests/tooling/r48/FailAction.java
@@ -14,20 +14,16 @@
  * limitations under the License.
  */
 
-package org.gradle.integtests.tooling.r47;
+package org.gradle.integtests.tooling.r48;
 
 import org.gradle.tooling.BuildAction;
 import org.gradle.tooling.BuildController;
 
 import java.io.Serializable;
 
-public class CustomBuildFinishedAction implements BuildAction<String>, Serializable {
-    // Tasks graph is already calculated and tasks executed. Action or model builders can access tasks results.
-
+public class FailAction implements BuildAction<String>, Serializable {
     @Override
     public String execute(BuildController controller) {
-        // Print something to verify it is after task execution
-        System.out.println("afterBuildAction");
-        return controller.getModel(CustomBuildFinishedModel.class).getValue();
+        throw new RuntimeException("actionFailure");
     }
 }
