+++ subprojects/tooling-api/src/main/java/org/gradle/tooling/ProjectConnection.java
@@ -152,10 +152,10 @@ public interface ProjectConnection {
      * Creates a builder for an executer which can be used to run actions in different phases of the build.
      * The actions are serialized into the build process and executed, then its result is serialized back to the caller.
      *
-     * <p>Requires Gradle 4.7 or later.
+     * <p>Requires Gradle 4.8 or later.
      *
      * @return The builder.
-     * @since 4.7
+     * @since 4.8
      */
     @Incubating
     PhasedBuildActionExecuter.Builder phasedAction();
