+++ subprojects/tooling-api/src/main/java/org/gradle/tooling/internal/consumer/connection/PhasedActionAwareConsumerConnection.java
@@ -38,7 +38,7 @@ import java.io.File;
 /**
  * An adapter for {@link InternalPhasedActionConnection}.
  *
- * <p>Used for providers >= 4.7.</p>
+ * <p>Used for providers >= 4.8.</p>
  */
 public class PhasedActionAwareConsumerConnection extends ParameterAcceptingConsumerConnection {
