+++ subprojects/tooling-api/src/main/java/org/gradle/tooling/internal/consumer/connection/UnsupportedOlderVersionConnection.java
@@ -68,7 +68,7 @@ public class UnsupportedOlderVersionConnection implements ConsumerConnection {
 
     @Override
     public void run(PhasedBuildAction phasedBuildAction, ConsumerOperationParameters operationParameters) {
-        throw Exceptions.unsupportedFeature(operationParameters.getEntryPointName(), version, "4.7");
+        throw Exceptions.unsupportedFeature(operationParameters.getEntryPointName(), version, "4.8");
     }
 
     public void runTests(TestExecutionRequest testExecutionRequest, ConsumerOperationParameters operationParameters) {
