+++ subprojects/tooling-api/src/main/java/org/gradle/tooling/internal/protocol/PhasedActionResultListener.java
@@ -21,14 +21,14 @@ package org.gradle.tooling.internal.protocol;
  *
  * A listener for results from each one of the actions in an {@link InternalPhasedAction}.
  *
- * @since 4.7
+ * @since 4.8
  */
 public interface PhasedActionResultListener extends InternalProtocolInterface {
 
     /**
      * Method called when result is available.
      *
-     * @since 4.7
+     * @since 4.8
      */
     void onResult(PhasedActionResult<?> result);
 }
