+++ subprojects/tooling-api/src/test/groovy/org/gradle/tooling/internal/consumer/connection/ParameterAcceptingConsumerConnectionTest.groovy
@@ -151,7 +151,7 @@ class ParameterAcceptingConsumerConnectionTest extends Specification {
 
         then:
         UnsupportedVersionException e = thrown()
-        e.message == 'The version of Gradle you are using (4.4) does not support the PhasedBuildActionExecuter API. Support for this is available in Gradle 4.7 and all later versions.'
+        e.message == 'The version of Gradle you are using (4.4) does not support the PhasedBuildActionExecuter API. Support for this is available in Gradle 4.8 and all later versions.'
     }
 
     interface TestModelBuilder extends ConnectionVersion4, ConfigurableConnection, InternalParameterAcceptingConnection, InternalTestExecutionConnection, StoppableConnection,
